requirements_minimal.txt
# 精简版机器人状态监控程序依赖包
# 只保留必要的核心功能依赖

# Web框架
Flask==2.2.2
Flask-SocketIO==5.3.6
Werkzeug==2.2.2

# WebSocket支持
python-socketio==5.11.0
python-engineio==4.8.0
bidict==0.22.1
simple-websocket==1.0.0

# 串口通信
pyserial==3.5

# 配置文件处理
PyYAML==6.0.1

# 系统信息获取
psutil==5.9.4
netifaces==0.11.0

# 基础依赖
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
blinker==1.5

# 树莓派GPIO（如果在树莓派上运行）
RPi.GPIO==0.7.1a4
lgpio==0.2.2.0

# 网络工具
requests==2.31.0
urllib3==1.26.12
certifi==2022.9.24
charset-normalizer==3.0.1
idna==3.3